create a subfolder two-factor-monitoring and move the files that use that subcomand
