+++ src/domain/usecases/user-monitoring/two-factor-monitoring/RunUserMonitoringReportUseCase.ts
@@ -6,8 +6,8 @@ import {
     UserMonitoringCountResponse,
     UsersOptions,
 } from "domain/entities/user-monitoring/UserMonitoring";
-import { MetadataRepository } from "domain/repositories/user-monitoring/MetadataRepository";
-import { ReportRepository } from "domain/repositories/user-monitoring/ReportRepository";
+import { MetadataRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/MetadataRepository";
+import { ReportRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/ReportRepository";
 import _ from "lodash";
 
 export class RunUserMonitoringReportUseCase {
