+++ src/domain/usecases/user-monitoring/two-factor-monitoring/RunUserMonitoringUserGroupsUseCase.ts
@@ -6,9 +6,9 @@ import {
     UserMonitoringCountResponse,
     UsersOptions,
 } from "domain/entities/user-monitoring/UserMonitoring";
-import { UserGroupRepository } from "domain/repositories/user-monitoring/UserGroupRepository";
-import { MetadataRepository } from "domain/repositories/user-monitoring/MetadataRepository";
-import { UserRepository } from "domain/repositories/user-monitoring/UserRepository";
+import { UserGroupRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/UserGroupRepository";
+import { MetadataRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/MetadataRepository";
+import { UserRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/UserRepository";
 import _ from "lodash";
 import log from "utils/log";
